packages/zeeker/pyproject.toml
[project]
name = "zeeker"
version = "0.6.0"
description = "Database management tool for Datasette-based systems"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "Ang Hou Fu", email = "houfu@outlook.sg"}]
keywords = ["datasette", "database", "sqlite", "data-management"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
]

dependencies = [
    "boto3>=1.38.32",
    "click>=8.2.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.2",
    "sqlite-utils>=3.38",
]

[project.optional-dependencies]
data = [
    "httpx>=0.27.0",
    "tenacity>=8.2.0",
]

[project.urls]
Homepage = "https://github.com/houfu/zeeker"
Documentation = "https://github.com/houfu/zeeker/blob/master/README.md"
Repository = "https://github.com/houfu/zeeker"
Issues = "https://github.com/houfu/zeeker/issues"

[project.scripts]
zeeker = "zeeker.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
